backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
sentence-transformers==2.5.1
chromadb==0.4.14
torch==2.2.1
numpy==1.26.4
hnswlib==0.7.0
PyJWT==2.8.0
requests==2.31.0